Применил аннотацию ConditionalOnProperty
