--- src/main/java/com/example/tt_nsk/controller/EnrollTournament_v2.java
@@ -1,15 +1,0 @@
-package com.example.tt_nsk.controller;
-
-import io.swagger.v3.oas.annotations.tags.Tag;
-import lombok.AllArgsConstructor;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@AllArgsConstructor
-@RequestMapping("/tournaments/enrollment")
-@Tag(name = "Контроллер, позволяющий регистрировать игроков на турниры")
-public class EnrollTournament_v2 {
-
-
-}
